requirements.txt
django
djangorestframework
datetime
meteomatics
whitenoise